# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Advancia Smart Wallets API
  version: "1.0.0"
  description: |
    REST API for the Advancia Smart Wallets platform — a fintech super-app
    with embedded smart wallets (ERC-4337), payments, healthcare, banking,
    loyalty, subscriptions, and admin management.

    ## Authentication
    Most endpoints require a JWT session token obtained from `/api/auth/*`.
    Pass it as a cookie (`session`) or `Authorization: Bearer <token>` header.

    Admin endpoints require a separate admin session via `/api/admin/login`.

    ## Versioning
    All endpoints are available under both `/api/*` and `/api/v1/*`.
    The `X-API-Version` response header indicates the active version.

    ## Rate Limiting
    Auth endpoints (`/api/auth/*`) are rate-limited to 20 requests/minute per IP.
    Exceeding the limit returns `429 Too Many Requests` with a `Retry-After` header.
  contact:
    name: Advancia Dev Team
    email: security@advancia.health
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://advancia-healthcare1.vercel.app
    description: Production (Vercel)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
    cronAuth:
      type: http
      scheme: bearer
      description: CRON_SECRET bearer token for cron endpoints
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "Invalid request body"
...
  responses:
...
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
security:
  - bearerAuth: []
  - cookieAuth: []

# ═══════════════════════════════════════════════════════════════
#  PATHS — grouped by tag / domain
# ═══════════════════════════════════════════════════════════════
paths:
...
  /api/transfers:
    get:
      tags: [Transactions]
      summary: List transfers (paginated)
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200": { description: "Transfer list with pagination" }
    post:
      tags: [Transactions]
      summary: Initiate a transfer
      responses:
        "201": { description: "Transfer initiated" }
        "400": { $ref: "#/components/responses/BadRequest" }
...
  /api/subscriptions:
    get:
      tags: [Subscriptions]
      summary: List user subscriptions (paginated)
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200": { description: "Paginated subscriptions" }
    post:
      tags: [Subscriptions]
      summary: Subscribe to a tier
      responses:
        "201": { description: "Subscription created" }
    patch:
      tags: [Subscriptions]
      summary: Cancel, pause, or resume subscription
      responses:
        "200": { description: "Subscription updated" }

  # ─── Social ───────────────────────────────────
...